refactor(search): move per-provider VEC tasks into parallel module (#349)
- Extract runSingleProviderVectorSearch and createProviderVectorSearchTask to
  hybrid-search-provider-parallel.ts with ProviderVectorSearchDeps injection
- HybridSearchEngine delegates via getProviderVectorSearchDeps; behavior unchanged
- Add design spec and implementation plan under docs/superpowers/
