+++ packages/memento-core/src/domains/search/algorithms/hybrid-search-provider-parallel.ts
@@ -1,10 +1,10 @@
 /**
- * Multi-provider VEC search: overall timeout, Promise.allSettled aggregation, diagnostics.
- * Extracted from HybridSearchEngine to reduce method size (Issue 315).
+ * Multi-provider VEC search: per-provider tasks, overall timeout, Promise.allSettled aggregation, diagnostics.
+ * Extracted from HybridSearchEngine (Issues #315, #349).
  */
 
 import { HYBRID_SEARCH } from '../../../shared/config/constants.js';
-import type { EmbeddingProvider } from '../../../shared/types/index.js';
+import type { EmbeddingProvider, MemoryType } from '../../../shared/types/index.js';
 import type { VectorSearchResult } from '../../memory/services/memory-embedding-service.js';
 
 export type VectorResultWithProvider = VectorSearchResult & { provider: string };
@@ -36,6 +36,145 @@ export type ProviderSearchExecutionSummary = {
 };
 
 export type SearchStepLogger = (searchId: string, step: string, data: unknown) => void;
+
+export type ProviderVectorSearchOptions = {
+  limit: number;
+  threshold: number;
+  types?: MemoryType[];
+  includeContent: boolean;
+};
+
+type VectorSearchRow = {
+  memory_id: string;
+  content: string;
+  type: string;
+  importance: number;
+  created_at: string;
+  similarity: number;
+};
+
+/** Injected from HybridSearchEngine so this module stays free of SearchError / engine imports. */
+export type ProviderVectorSearchDeps = {
+  generateQueryVector: (
+    query: string,
+    searchId: string,
+    preferredProvider: EmbeddingProvider
+  ) => Promise<{ embedding: number[]; actualProvider: EmbeddingProvider }>;
+  vectorSearch: (
+    vector: number[],
+    options: ProviderVectorSearchOptions,
+    provider?: string
+  ) => Promise<VectorSearchRow[]>;
+  logSearchStep: SearchStepLogger;
+};
+
+export async function runSingleProviderVectorSearch(
+  deps: ProviderVectorSearchDeps,
+  provider: EmbeddingProvider,
+  query: string,
+  searchOptions: ProviderVectorSearchOptions,
+  searchId: string,
+  providerStartTime: bigint,
+  timeoutMeta: { queryEmbeddingProvider?: EmbeddingProvider; tfidfQueryEmbeddingFallback?: boolean }
+): Promise<ProviderVectorRaceResult> {
+  try {
+    const { embedding, actualProvider } = await deps.generateQueryVector(query, searchId, provider);
+    timeoutMeta.queryEmbeddingProvider = actualProvider;
+    timeoutMeta.tfidfQueryEmbeddingFallback = actualProvider === 'tfidf' && provider !== 'tfidf';
+    if (actualProvider !== provider) {
+      deps.logSearchStep(searchId, `VEC 검색 스킵 (provider 불일치: 요청=${provider}, 실제=${actualProvider})`, {
+        provider,
+        actualProvider
+      });
+      const providerTime = Number(process.hrtime.bigint() - providerStartTime) / 1_000_000;
+      return {
+        provider,
+        results: [] as VectorResultWithProvider[],
+        success: true,
+        timeMs: providerTime,
+        error: null,
+        queryEmbeddingProvider: actualProvider,
+        tfidfQueryEmbeddingFallback: timeoutMeta.tfidfQueryEmbeddingFallback
+      };
+    }
+    const vecResults = await deps.vectorSearch(embedding, searchOptions, provider);
+    const providerTime = Number(process.hrtime.bigint() - providerStartTime) / 1_000_000;
+    return {
+      provider,
+      results: vecResults.map(result => ({
+        id: result.memory_id,
+        content: result.content,
+        type: result.type,
+        importance: result.importance,
+        created_at: result.created_at,
+        pinned: false,
+        score: result.similarity,
+        similarity: result.similarity,
+        provider
+      })),
+      success: true,
+      timeMs: providerTime,
+      error: null,
+      queryEmbeddingProvider: actualProvider,
+      tfidfQueryEmbeddingFallback: false
+    };
+  } catch (error) {
+    const providerTime = Number(process.hrtime.bigint() - providerStartTime) / 1_000_000;
+    const errorMessage = error instanceof Error ? error.message : String(error);
+    deps.logSearchStep(searchId, `VEC 벡터 검색 실패 - ${provider}`, {
+      provider,
+      error: errorMessage,
+      timeMs: providerTime
+    });
+    return {
+      provider,
+      results: [] as VectorResultWithProvider[],
+      success: false,
+      timeMs: providerTime,
+      error: errorMessage
+    };
+  }
+}
+
+export function createProviderVectorSearchTask(
+  deps: ProviderVectorSearchDeps,
+  provider: EmbeddingProvider,
+  query: string,
+  searchOptions: ProviderVectorSearchOptions,
+  searchId: string
+): Promise<ProviderVectorRaceResult> {
+  const providerStartTime = process.hrtime.bigint();
+  const timeoutMeta: {
+    queryEmbeddingProvider?: EmbeddingProvider;
+    tfidfQueryEmbeddingFallback?: boolean;
+  } = {};
+
+  const timeoutPromise = new Promise<ProviderVectorRaceResult>(resolve => {
+    setTimeout(() => {
+      const providerTime = Number(process.hrtime.bigint() - providerStartTime) / 1_000_000;
+      resolve({
+        provider,
+        results: [] as VectorResultWithProvider[],
+        success: false,
+        timeMs: providerTime,
+        error: `Provider 검색 타임아웃 (${HYBRID_SEARCH.PROVIDER_SEARCH_TIMEOUT_MS}ms 초과)`,
+        queryEmbeddingProvider: timeoutMeta.queryEmbeddingProvider,
+        tfidfQueryEmbeddingFallback: timeoutMeta.tfidfQueryEmbeddingFallback
+      });
+    }, HYBRID_SEARCH.PROVIDER_SEARCH_TIMEOUT_MS);
+  });
+
+  const searchTask = runSingleProviderVectorSearch(
+    deps,
+    provider,
+    query,
+    searchOptions,
+    searchId,
+    providerStartTime,
+    timeoutMeta
+  );
+  return Promise.race([searchTask, timeoutPromise]);
+}
 
 export async function executeProviderSearchesWithOverallTimeout(
   searchPromises: Promise<ProviderVectorRaceResult>[],
